+++ packages/editor/editor-core/src/__tests__/unit/plugins/media/index.ts
@@ -39,7 +39,6 @@ import listPlugin from '../../../../plugins/lists';
 import codeBlockPlugin from '../../../../plugins/code-block';
 import rulePlugin from '../../../../plugins/rule';
 import tablePlugin from '../../../../plugins/table';
-import pickerFacadeLoader from '../../../../plugins/media/picker-facade-loader';
 import { insertMediaAsMediaSingle } from '../../../../plugins/media/utils/media-single';
 
 const stateManager = new DefaultMediaStateManager();
@@ -112,11 +111,16 @@ describe('Media plugin', () => {
     return mediaNodeWithPos!.getPos();
   };
 
-  const waitForMediaPickerReady = async (pluginState: MediaPluginState) =>
-    Promise.all([
-      new Promise(resolve => pluginState.subscribe(resolve)),
-      pickerFacadeLoader(),
-    ]);
+  const waitForMediaPickerReady = (pluginState: MediaPluginState) =>
+    new Promise(resolve => pluginState.subscribe(resolve));
+
+  const waitForAllPickersInitialised = async (
+    pluginState: MediaPluginState,
+  ) => {
+    while (pluginState.pickers.length < 4) {
+      await waitForMediaPickerReady(pluginState);
+    }
+  };
 
   afterAll(() => {
     providerFactory.destroy();
@@ -133,7 +137,8 @@ describe('Media plugin', () => {
     const provider = await mediaProvider;
     await provider.uploadContext;
 
-    await waitForMediaPickerReady(pluginState);
+    await waitForAllPickersInitialised(pluginState);
+
     expect(typeof pluginState.binaryPicker!).toBe('object');
 
     pluginState.binaryPicker!.upload = jest.fn();
@@ -625,15 +630,10 @@ describe('Media plugin', () => {
     const { pluginState } = editor(doc(h1('text{<>}')));
     expect(pluginState.pickers.length).toBe(0);
 
-    const mediaProvider1 = getFreshMediaProvider();
-    await pluginState.setMediaProvider(mediaProvider1);
-    const mediaProvider2 = getFreshMediaProvider();
-    await pluginState.setMediaProvider(mediaProvider2);
+    await getFreshMediaProvider();
+    await getFreshMediaProvider();
 
-    const resolvedMediaProvider1 = await mediaProvider1;
-    const resolvedMediaProvider2 = await mediaProvider2;
-    await resolvedMediaProvider1.uploadContext;
-    await resolvedMediaProvider2.uploadContext;
+    await waitForAllPickersInitialised(pluginState);
 
     expect(pluginState.pickers.length).toBe(4);
   });
@@ -642,17 +642,17 @@ describe('Media plugin', () => {
     const { pluginState } = editor(doc(h1('text{<>}')));
     expect(pluginState.pickers.length).toBe(0);
 
-    const mediaProvider1 = getFreshMediaProvider();
-    await pluginState.setMediaProvider(mediaProvider1);
-    const resolvedMediaProvider1 = await mediaProvider1;
-    await resolvedMediaProvider1.uploadContext;
+    await getFreshMediaProvider();
+
+    await waitForAllPickersInitialised(pluginState);
+
     const pickersAfterMediaProvider1 = pluginState.pickers;
     expect(pickersAfterMediaProvider1.length).toBe(4);
 
-    const mediaProvider2 = getFreshMediaProvider();
-    await pluginState.setMediaProvider(mediaProvider2);
-    const resolvedMediaProvider2 = await mediaProvider2;
-    await resolvedMediaProvider2.uploadContext;
+    await getFreshMediaProvider();
+
+    await waitForAllPickersInitialised(pluginState);
+
     const pickersAfterMediaProvider2 = pluginState.pickers;
 
     expect(pickersAfterMediaProvider1).toHaveLength(
@@ -700,7 +700,7 @@ describe('Media plugin', () => {
     const provider = await mediaProvider;
     await provider.uploadContext;
     await provider.viewContext;
-    await waitForMediaPickerReady(pluginState);
+    await waitForAllPickersInitialised(pluginState);
     expect(typeof pluginState.binaryPicker!).toBe('object');
 
     const testFileData = {
@@ -743,7 +743,7 @@ describe('Media plugin', () => {
     const provider = await mediaProvider;
     await provider.uploadContext;
     await provider.viewContext;
-    await waitForMediaPickerReady(pluginState);
+    await waitForAllPickersInitialised(pluginState);
     expect(typeof pluginState.binaryPicker!).toBe('object');
 
     const testFileData = {
@@ -781,7 +781,7 @@ describe('Media plugin', () => {
     const provider = await mediaProvider;
     await provider.uploadContext;
     await provider.viewContext;
-    await waitForMediaPickerReady(pluginState);
+    await waitForAllPickersInitialised(pluginState);
     expect(typeof pluginState.binaryPicker!).toBe('object');
 
     const testFileData = {
@@ -817,7 +817,8 @@ describe('Media plugin', () => {
     const provider = await mediaProvider;
     await provider.uploadContext;
     await provider.viewContext;
-    await waitForMediaPickerReady(pluginState);
+    await waitForAllPickersInitialised(pluginState);
+
     expect(typeof pluginState.binaryPicker!).toBe('object');
 
     const testFileData = {