+++ packages/editor/renderer/src/__tests__/unit/react/nodes/mediaGroup.tsx
@@ -128,4 +128,17 @@ describe('MediaGroup', () => {
 
     mediaGroup.unmount();
   });
+
+  it('should send useInlinePlayer: false to the Media', () => {
+    const mediaGroup = mount(
+      <MediaGroup>
+        <Media
+          id={imageFileId.id}
+          type={imageFileId.mediaItemType}
+          collection={imageFileId.collectionName}
+        />
+      </MediaGroup>,
+    );
+    expect(mediaGroup.find(Media).prop('useInlinePlayer')).toBe(false);
+  });
 });