+++ packages/elements/user-picker/src/components/MultiValue.tsx
@@ -1,10 +1,11 @@
 import Tag from '@atlaskit/tag';
 import * as React from 'react';
 import { FormattedMessage } from 'react-intl';
-import { Option } from '../types';
+import { Option, UserPickerProps } from '../types';
+import { AddOptionAvatar } from './AddOptionAvatar';
 import { messages } from './i18n';
 import { SizeableAvatar } from './SizeableAvatar';
-import { getAvatarUrl } from './utils';
+import { getAvatarUrl, isEmail } from './utils';
 
 export const scrollToValue = (
   valueContainer: HTMLDivElement,
@@ -25,7 +26,7 @@ type Props = {
   data: Option;
   innerProps: any;
   removeProps: { onClick: Function };
-  selectProps: { isDisabled: boolean };
+  selectProps: UserPickerProps;
 };
 
 export class MultiValue extends React.Component<Props> {
@@ -73,6 +74,29 @@ export class MultiValue extends React.Component<Props> {
     );
   }
 
+  getElemBefore = () => {
+    const {
+      data: { data, label },
+      selectProps,
+    } = this.props;
+    if (isEmail(data)) {
+      return selectProps.emailLabel ? (
+        <AddOptionAvatar size="small" label={selectProps.emailLabel} />
+      ) : (
+        <FormattedMessage {...messages.addEmail}>
+          {label => <AddOptionAvatar size="small" label={label as string} />}
+        </FormattedMessage>
+      );
+    }
+    return (
+      <SizeableAvatar
+        appearance="multi"
+        src={getAvatarUrl(data)}
+        name={label}
+      />
+    );
+  };
+
   render() {
     const {
       data: { label, data },
@@ -90,13 +114,7 @@ export class MultiValue extends React.Component<Props> {
               {...innerProps}
               appearance="rounded"
               text={label}
-              elemBefore={
-                <SizeableAvatar
-                  appearance="multi"
-                  src={getAvatarUrl(data)}
-                  name={label}
-                />
-              }
+              elemBefore={this.getElemBefore()}
               removeButtonText={data.fixed || isDisabled ? undefined : remove}
               onAfterRemoveAction={onRemove}
               color={isFocused ? 'blueLight' : undefined}