+++ packages/media/media-picker/examples/11-dropzone-i18n.tsx
@@ -41,13 +41,13 @@ class DropzoneWrapper extends Component<{}, DropzoneWrapperState> {
     }
   }
 
-  createMediaPicker(reactContext: any) {
+  async createMediaPicker(reactContext: any) {
     if (!this.dropzoneRef.current) {
       return;
     }
 
     const context = createUploadContext();
-    const dropzone = MediaPicker('dropzone', context, {
+    const dropzone = await MediaPicker('dropzone', context, {
       container: this.dropzoneRef.current,
       uploadParams: {
         collection: defaultMediaPickerCollectionName,
@@ -58,8 +58,8 @@ class DropzoneWrapper extends Component<{}, DropzoneWrapperState> {
     dropzone.activate();
   }
 
-  componentDidMount = () => {
-    this.createMediaPicker(this.context);
+  componentDidMount = async () => {
+    await this.createMediaPicker(this.context);
   };
 
   render() {