+++ packages/media/media-picker/src/components/__tests__/unit/browser.spec.ts
@@ -1,7 +1,8 @@
 jest.mock('../../../service/newUploadServiceImpl');
 
 import { Context, ContextFactory } from '@atlaskit/media-core';
-import { Browser, BrowserConfig } from '../../browser';
+import { Browser, BrowserConfig } from '../../types';
+import { BrowserImpl } from '../../browser';
 
 describe('Browser', () => {
   let browser: Browser | undefined;
@@ -27,13 +28,13 @@ describe('Browser', () => {
 
   it('should append the input to the body', () => {
     const inputsBefore = document.querySelectorAll('input[type=file]');
-    browser = new Browser(context, browseConfig);
+    browser = new BrowserImpl(context, browseConfig);
     const inputsAfter = document.querySelectorAll('input[type=file]');
     expect(inputsAfter.length).toBeGreaterThan(inputsBefore.length);
   });
 
   it('should remove the input from the body', () => {
-    browser = new Browser(context, browseConfig);
+    browser = new BrowserImpl(context, browseConfig);
     const inputsBefore = document.querySelectorAll('input[type=file]');
     browser.teardown();
     const inputsAfter = document.querySelectorAll('input[type=file]');
@@ -48,7 +49,7 @@ describe('Browser', () => {
     input.addEventListener = addEventListener;
     input.removeEventListener = removeEventListener;
     document.createElement = jest.fn().mockReturnValue(input);
-    browser = new Browser(context, browseConfig);
+    browser = new BrowserImpl(context, browseConfig);
     expect(addEventListener).toHaveBeenCalledTimes(1);
     expect(addEventListener.mock.calls[0][0]).toEqual('change');
 
@@ -60,7 +61,7 @@ describe('Browser', () => {
   it('should add upload files when user picks some', () => {
     const input = document.createElement('input');
     document.createElement = jest.fn().mockReturnValue(input);
-    browser = new Browser(context, browseConfig);
+    browser = new BrowserImpl(context, browseConfig);
 
     const spy = jest.spyOn(browser['uploadService'], 'addFiles');
     input.dispatchEvent(new Event('change'));