+++ packages/media/media-picker/src/popup/components/app.tsx
@@ -12,13 +12,10 @@ import {
 
 import { ServiceName, State } from '../domain';
 
-import {
-  BinaryUploader as MpBinary,
-  Browser as MpBrowser,
-  Dropzone as MpDropzone,
-  UploadParams,
-  PopupConfig,
-} from '../..';
+import { BinaryUploaderImpl as MpBinary } from '../../components/binary';
+import { BrowserImpl as MpBrowser } from '../../components/browser';
+import { DropzoneImpl as MpDropzone } from '../../components/dropzone';
+import { UploadParams, PopupConfig } from '../..';
 
 /* Components */
 import Footer from './footer/footer';
@@ -44,7 +41,6 @@ import { fileUploadError } from '../actions/fileUploadError';
 import { dropzoneDropIn } from '../actions/dropzoneDropIn';
 import { dropzoneDragIn } from '../actions/dropzoneDragIn';
 import { dropzoneDragOut } from '../actions/dropzoneDragOut';
-import { MediaPicker } from '../..';
 import PassContext from './passContext';
 import {
   UploadsStartEventPayload,
@@ -58,7 +54,7 @@ import { MediaPickerPopupWrapper, SidebarWrapper, ViewWrapper } from './styled';
 import {
   DropzoneDragEnterEventPayload,
   DropzoneDragLeaveEventPayload,
-} from '../../components/dropzone';
+} from '../../components/types';
 
 export interface AppStateProps {
   readonly selectedServiceName: ServiceName;
@@ -137,7 +133,7 @@ export class App extends Component<AppProps, AppState> {
       cacheSize: tenantContext.config.cacheSize,
     });
 
-    this.mpBrowser = MediaPicker('browser', context, {
+    this.mpBrowser = new MpBrowser(context, {
       uploadParams: tenantUploadParams,
       shouldCopyFileToRecents: false,
       multiple: true,
@@ -149,7 +145,7 @@ export class App extends Component<AppProps, AppState> {
     this.mpBrowser.on('upload-end', onUploadEnd);
     this.mpBrowser.on('upload-error', onUploadError);
 
-    this.mpDropzone = MediaPicker('dropzone', context, {
+    this.mpDropzone = new MpDropzone(context, {
       uploadParams: tenantUploadParams,
       shouldCopyFileToRecents: false,
       headless: true,
@@ -163,7 +159,7 @@ export class App extends Component<AppProps, AppState> {
     this.mpDropzone.on('upload-end', onUploadEnd);
     this.mpDropzone.on('upload-error', onUploadError);
 
-    this.mpBinary = MediaPicker('binary', context, {
+    this.mpBinary = new MpBinary(context, {
       uploadParams: tenantUploadParams,
       shouldCopyFileToRecents: false,
     });