+++ packages/media/media-picker/src/popup/middleware/importFiles.tsx
@@ -23,7 +23,7 @@ import { WsProvider } from '../tools/websocket/wsProvider';
 import { WsConnectionHolder } from '../tools/websocket/wsConnectionHolder';
 import { RemoteUploadActivity } from '../tools/websocket/upload/remoteUploadActivity';
 import { MediaFile, copyMediaFileForUpload } from '../../domain/file';
-import { PopupUploadEventEmitter } from '../../components/popup';
+import { PopupUploadEventEmitter } from '../../components/types';
 import { sendUploadEvent } from '../actions/sendUploadEvent';
 import { setUpfrontIdDeferred } from '../actions/setUpfrontIdDeferred';
 import { WsNotifyMetadata } from '../tools/websocket/wsMessageData';